+++ spec/acceptance/api/v2/contacts/merges_spec.rb
@@ -85,6 +85,7 @@ resource 'Contacts > Merges' do
       donor_accounts
       last_six_donations
       people
+      primary_person
       tasks
     )
   end