+++ spec/acceptance/api/v2/contacts_spec.rb
@@ -107,6 +107,7 @@ resource 'Contacts' do
       donor_accounts
       last_six_donations
       people
+      primary_person
       tasks
     )
   end